+++ docs/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next
 
+* Add search filters based on character location on dim (is:inleftchar / inmiddlechar / inrightchar) and for vault (is:invault) and current/last logged character (incurrentchar), that is marked with a yellow triangle.
 * Fixed a bug where the "Restore Old Versions" tool wouldn't actually let you see and restore old versions.
 
 # 4.49.1 (2018-04-23)